Fix warning

+++ INSElectronicProgramGuideLayout/INSElectronicProgramGuideLayout.m
@@ -687,7 +687,7 @@ NSUInteger const INSEPGLayoutMinBackgroundZ = 0.0;
     NSIndexPath *hourHeaderBackgroundIndexPath = [NSIndexPath indexPathForRow:0 inSection:0];
     UICollectionViewLayoutAttributes *hourHeaderBackgroundAttributes = [self layoutAttributesForDecorationViewAtIndexPath:hourHeaderBackgroundIndexPath ofKind:INSEPGLayoutElementKindHourHeaderBackground withItemCache:self.hourHeaderBackgroundAttributes];
     // Frame
-    CGFloat hourHeaderBackgroundHeight = (self.hourHeaderHeight + ((self.collectionView.contentOffset.y < 0.0) ? fabsf(self.collectionView.contentOffset.y) : 0.0)) - self.collectionView.contentInset.top;
+    CGFloat hourHeaderBackgroundHeight = (self.hourHeaderHeight + ((self.collectionView.contentOffset.y < 0.0) ? fabs(self.collectionView.contentOffset.y) : 0.0)) - self.collectionView.contentInset.top;
 
     if (!self.shouldResizeStickyHeaders || self.hourHeaderHeight >= hourHeaderBackgroundHeight) {
         hourHeaderBackgroundHeight = self.hourHeaderHeight;
